Add a button on the app bar to go on the setting activity
