+++ Amio/app/src/main/java/eu/telecomnancy/amio/SettingsActivity.java
@@ -11,8 +11,7 @@ import eu.telecomnancy.amio.iotlab.PollingService;
 
 public class SettingsActivity extends AppCompatActivity {
 
-    // TODO: move to main activity once created
-    private Intent pollingServiceIntent;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -26,18 +25,6 @@ public class SettingsActivity extends AppCompatActivity {
         if (actionBar != null) {
             actionBar.setDisplayHomeAsUpEnabled(true);
         }
-
-        // TODO: move to main activity once created
-        pollingServiceIntent = new Intent(this, PollingService.class);
-        startService(pollingServiceIntent);
-    }
-
-    @Override
-    protected void onDestroy() {
-        super.onDestroy();
-
-        // TODO: move to main activity once created
-        stopService(pollingServiceIntent);
     }
 
     public static class SettingsFragment extends PreferenceFragmentCompat {
